+++ scheduler/scheduler/src/main/kotlin/uk/ac/ebi/scheduler/scheduling/DailyScheduler.kt
@@ -1,51 +1,50 @@
 package uk.ac.ebi.scheduler.scheduling
 
-import ebi.ac.uk.base.ifTrue
-import uk.ac.ebi.scheduler.pmc.importer.domain.PmcLoaderService
 import org.springframework.scheduling.annotation.Scheduled
 import uk.ac.ebi.scheduler.common.properties.DailyScheduling
-import uk.ac.ebi.scheduler.exporter.domain.ExporterTrigger
+import uk.ac.ebi.scheduler.pmc.exporter.domain.ExporterTrigger
 import uk.ac.ebi.scheduler.pmc.importer.DEFAULT_FOLDER
+import uk.ac.ebi.scheduler.pmc.importer.domain.PmcLoaderService
 import uk.ac.ebi.scheduler.releaser.domain.SubmissionReleaserTrigger
 
 internal class DailyScheduler(
     private val dailyScheduling: DailyScheduling,
     private val exporterTrigger: ExporterTrigger,
     private val pmcLoaderService: PmcLoaderService,
-    private val submissionReleaserTrigger: SubmissionReleaserTrigger
+    private val submissionReleaserTrigger: SubmissionReleaserTrigger,
 ) {
     @Scheduled(cron = "0 0 1 * * *")
     fun releaseSubmissions() {
-        dailyScheduling.releaser.ifTrue { submissionReleaserTrigger.triggerSubmissionReleaser() }
+        if (dailyScheduling.releaser) submissionReleaserTrigger.triggerSubmissionReleaser()
     }
 
     @Scheduled(cron = "0 0 6 * * *")
     fun loadPmc() {
-        dailyScheduling.pmc.ifTrue { pmcLoaderService.loadFile(DEFAULT_FOLDER) }
+        if (dailyScheduling.pmc) pmcLoaderService.loadFile(DEFAULT_FOLDER)
     }
 
     @Scheduled(cron = "0 0 7 * * *")
     fun processPmc() {
-        dailyScheduling.pmc.ifTrue { pmcLoaderService.triggerProcessor() }
+        if (dailyScheduling.pmc) pmcLoaderService.triggerProcessor()
     }
 
     @Scheduled(cron = "0 0 8 * * *")
     fun submitPmc() {
-        dailyScheduling.pmc.ifTrue { pmcLoaderService.triggerSubmitter() }
+        if (dailyScheduling.pmc) pmcLoaderService.triggerSubmitter()
     }
 
     @Scheduled(cron = "0 0 10 * * *")
     fun notifySubmissionRelease() {
-        dailyScheduling.notifier.ifTrue { submissionReleaserTrigger.triggerSubmissionReleaseNotifier() }
+        if (dailyScheduling.notifier) submissionReleaserTrigger.triggerSubmissionReleaseNotifier()
     }
 
     @Scheduled(cron = "0 0 20 * * *")
     fun exportPmcSubmissions() {
-        dailyScheduling.pmc.ifTrue { exporterTrigger.triggerPmcExport() }
+        if (dailyScheduling.pmc) exporterTrigger.triggerPmcExport()
     }
 
     @Scheduled(cron = "0 0 21 * * *")
     fun exportPublicSubmissions() {
-        dailyScheduling.exporter.ifTrue { exporterTrigger.triggerPublicExport() }
+        if (dailyScheduling.exporter) exporterTrigger.triggerPublicExport()
     }
 }
