--- scheduler/tasks/exporter-task/src/main/kotlin/uk/ac/ebi/scheduler/exporter/ExporterApp.kt
@@ -1,41 +1,0 @@
-package uk.ac.ebi.scheduler.exporter
-
-import ac.uk.ebi.scheduler.properties.ExporterMode.PMC
-import ac.uk.ebi.scheduler.properties.ExporterMode.PUBLIC_ONLY
-import org.springframework.boot.CommandLineRunner
-import org.springframework.boot.autoconfigure.SpringBootApplication
-import org.springframework.boot.context.properties.EnableConfigurationProperties
-import org.springframework.boot.runApplication
-import org.springframework.context.ApplicationContext
-import org.springframework.context.ApplicationContextAware
-import org.springframework.data.mongodb.repository.config.EnableMongoRepositories
-import uk.ac.ebi.scheduler.exporter.config.ApplicationProperties
-import uk.ac.ebi.scheduler.exporter.service.ExporterService
-
-@SpringBootApplication
-@EnableMongoRepositories
-@EnableConfigurationProperties
-class ExporterApp
-
-@Suppress("SpreadOperator")
-fun main(args: Array<String>) {
-    runApplication<ExporterApp>(*args)
-}
-
-class ExporterExecutor(
-    private val exporterService: ExporterService,
-    private val appProperties: ApplicationProperties
-) : CommandLineRunner, ApplicationContextAware {
-    private lateinit var context: ApplicationContext
-
-    override fun run(vararg args: String?) {
-        when (appProperties.mode) {
-            PMC -> exporterService.exportPmc()
-            PUBLIC_ONLY -> exporterService.exportPublicOnly()
-        }
-    }
-
-    override fun setApplicationContext(applicationContext: ApplicationContext) {
-        this.context = applicationContext
-    }
-}
