--- scheduler/tasks/exporter-task/src/main/kotlin/uk/ac/ebi/scheduler/exporter/model/PmcData.kt
@@ -1,6 +1,0 @@
-package uk.ac.ebi.scheduler.exporter.model
-
-data class PmcData(
-    val accNo: String,
-    val title: String?
-)
