+++ scheduler/tasks/exporter-task/src/main/kotlin/uk/ac/ebi/scheduler/pmc/exporter/config/ApplicationProperties.kt
@@ -1,0 +1,34 @@
+package uk.ac.ebi.scheduler.pmc.exporter.config
+
+import ac.uk.ebi.scheduler.properties.ExporterMode
+import org.springframework.boot.context.properties.ConfigurationProperties
+import org.springframework.boot.context.properties.NestedConfigurationProperty
+import org.springframework.stereotype.Component
+
+@ConfigurationProperties(prefix = "app")
+@Component
+class ApplicationProperties {
+    lateinit var fileName: String
+    lateinit var outputPath: String
+    lateinit var mode: ExporterMode
+    lateinit var tmpFilesPath: String
+
+    @NestedConfigurationProperty
+    var bioStudies: BioStudies = BioStudies()
+
+    @NestedConfigurationProperty
+    var ftp: Ftp = Ftp()
+}
+
+class BioStudies {
+    lateinit var url: String
+    lateinit var user: String
+    lateinit var password: String
+}
+
+class Ftp {
+    lateinit var host: String
+    lateinit var user: String
+    lateinit var password: String
+    lateinit var port: Number
+}
