+++ scheduler/tasks/exporter-task/src/main/kotlin/uk/ac/ebi/scheduler/pmc/exporter/model/PmcData.kt
@@ -1,0 +1,6 @@
+package uk.ac.ebi.scheduler.pmc.exporter.model
+
+data class PmcData(
+    val accNo: String,
+    val title: String?,
+)
