+++ scheduler/tasks/exporter-task/src/main/kotlin/uk/ac/ebi/scheduler/pmc/exporter/service/ExporterService.kt
@@ -1,0 +1,16 @@
+package uk.ac.ebi.scheduler.pmc.exporter.service
+
+import kotlinx.coroutines.runBlocking
+
+class ExporterService(
+    private val pmcExporterService: PmcExporterService,
+    private val publicOnlyExporterService: PublicOnlyExporterService,
+) {
+    fun exportPmc() {
+        runBlocking { pmcExporterService.exportPmcLinks() }
+    }
+
+    fun exportPublicOnly() {
+        publicOnlyExporterService.exportPublicSubmissions()
+    }
+}
